requirements.txt
allure-pytest==2.3.3b1
pytest==3.6.0
selenium==3.12.0
